Set-up next-translate paths
