+++ locales/en/common.json
@@ -1,0 +1,1 @@
+{}
