+++ src/pages/[chapterId].js
@@ -1,15 +1,15 @@
 // @ts-nocheck
 import I18nProvider from 'next-translate/I18nProvider'
 import React from 'react'
-import C, * as _rest from '../../../src/pages_/[chapterId]'
-import ns0 from '../../../locales/ar/common.json'
+import C, * as _rest from '../../src/pages_/[chapterId]'
+import ns0 from '../.././locales/en/common.json'
 
 const namespaces = { 'common': ns0 }
 
 export default function Page(p){
   return (
     <I18nProvider 
-      lang="ar" 
+      lang="en" 
       namespaces={namespaces}  
       internals={{"defaultLanguage":"en","isStaticMode":true}}
     >
@@ -21,11 +21,11 @@ export default function Page(p){
 Page = Object.assign(Page, { ...C })
 
 if(C && C.getInitialProps) {
-  Page.getInitialProps = ctx => C.getInitialProps({ ...ctx, lang: 'ar'})
+  Page.getInitialProps = ctx => C.getInitialProps({ ...ctx, lang: 'en'})
 }
 
-export const getStaticProps = ctx => _rest.getStaticProps({ ...ctx, lang: 'ar' })
-export const getStaticPaths = ctx => _rest.getStaticPaths({ ...ctx, lang: 'ar' })
+export const getStaticProps = ctx => _rest.getStaticProps({ ...ctx, lang: 'en' })
+export const getStaticPaths = ctx => _rest.getStaticPaths({ ...ctx, lang: 'en' })
 
 
-export * from '../../../src/pages_/[chapterId]'
+export * from '../../src/pages_/[chapterId]'
