+++ src/pages/ar/index.js
@@ -1,16 +1,16 @@
 // @ts-nocheck
 import I18nProvider from 'next-translate/I18nProvider'
 import React from 'react'
-import C, * as _rest from '../../src/pages_'
-import ns0 from '../../locales/en/common.json'
-import ns1 from '../../locales/en/home.json'
+import C, * as _rest from '../../../src/pages_'
+import ns0 from '../../.././locales/ar/common.json'
+import ns1 from '../../.././locales/ar/home.json'
 
 const namespaces = { 'common': ns0, 'home': ns1 }
 
 export default function Page(p){
   return (
     <I18nProvider 
-      lang="en" 
+      lang="ar" 
       namespaces={namespaces}  
       internals={{"defaultLanguage":"en","isStaticMode":true}}
     >
@@ -22,11 +22,11 @@ export default function Page(p){
 Page = Object.assign(Page, { ...C })
 
 if(C && C.getInitialProps) {
-  Page.getInitialProps = ctx => C.getInitialProps({ ...ctx, lang: 'en'})
+  Page.getInitialProps = ctx => C.getInitialProps({ ...ctx, lang: 'ar'})
 }
 
-export const getStaticProps = ctx => _rest.getStaticProps({ ...ctx, lang: 'en' })
+export const getStaticProps = ctx => _rest.getStaticProps({ ...ctx, lang: 'ar' })
 
 
 
-export * from '../../src/pages_'
+export * from '../../../src/pages_'
